requirements.txt
fastapi
uvicorn
httpx
pydantic
pyyaml
tiktoken
pyjwt
bcrypt
python-multipart